Mini refactoring of names
