+++ components/poll.vue
